Fixed #90.

+++ leads_gui/accelerometer.py
@@ -34,7 +34,7 @@ class GForceMeter(TextBased, VariableControlled):
                  fg_color: _Color | None = None,
                  hover_color: _Color | None = None,
                  bg_color: _Color | None = None,
-                 corner_radius: int | None = None) -> None:
+                 corner_radius: float | None = None) -> None:
         TextBased.__init__(self, master, theme_key, width, height, font, text_color, fg_color, hover_color, bg_color,
                            corner_radius)
         VariableControlled.__init__(self, variable if variable else GForceVar(master))
@@ -69,7 +69,7 @@ class SpeedTrendMeter(TextBased, VariableControlled):
                  fg_color: _Color | None = None,
                  hover_color: _Color | None = None,
                  bg_color: _Color | None = None,
-                 corner_radius: int | None = None) -> None:
+                 corner_radius: float | None = None) -> None:
         TextBased.__init__(self, master, theme_key, width, height, font, text_color, fg_color, hover_color, bg_color,
                            corner_radius)
         VariableControlled.__init__(self, variable if variable else _DoubleVar(master))
